:tada: Stream output git
https://git-scm.com/book/gr/v2/Git-Internals-Transfer-Protocols#_uploading_data

```
Each line starts with a 4-character hex value specifying how long the rest of the line is. Your first line starts with 005b, which is hexadecimal for 91, meaning that 91 bytes remain on that line. The next line starts with 003e, which is 62, so you read the remaining 62 bytes. The next line is 0000, meaning the server is done with its references listing.
```